Feature/346 jvm metric save to elastic search test success

+++ apm-collector/apm-collector-stream/src/main/java/org/skywalking/apm/collector/stream/worker/impl/data/operate/AddOperation.java
@@ -15,7 +15,7 @@ public class AddOperation implements Operation {
         return newValue + oldValue;
     }
 
-    @Override public Float operate(Float newValue, Float oldValue) {
+    @Override public Double operate(Double newValue, Double oldValue) {
         return newValue + oldValue;
     }
